refact(provenance): factor out counts display component
